+++ src/test/java/org/cicirello/search/restarts/MultistartSearchTests.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2020  Vincent A. Cicirello
+ * Copyright (C) 2002-2021 Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  * 
@@ -515,7 +515,8 @@ public class MultistartSearchTests {
 					if (elapsed >= stopAtEval) {
 						elapsed = stopAtEval;
 						tracker.stop();
-						tracker.setFoundBest();
+						// replaced deprecated call to setFoundBest()
+						tracker.update(1, new TestObject(), true);
 						c = 1;
 					}
 					break;
@@ -528,7 +529,8 @@ public class MultistartSearchTests {
 				case 2:
 					if (elapsed >= findBestAtEval) {
 						elapsed = findBestAtEval;
-						tracker.setFoundBest();
+						// replaced deprecated call to setFoundBest()
+						tracker.update(1, new TestObject(), true);
 						c = 1;
 					}
 					break;